+++ hosts/sigma/hardware-configuration.nix
@@ -8,28 +8,24 @@
   boot.extraModulePackages = [ ];
 
   disko.devices = {
-    disk.nvme0n1 = {
+    disk.primary = {
       type = "disk";
       device = "/dev/nvme0n1";
       content = {
-        type = "table";
-        format = "gpt";
-        partitions = [
-          {
-            name = "ESP";
-            start = "1MiB";
-            end = "512MiB";
-            bootable = true;
+        type = "gpt";
+        partitions = {
+          esp = {
+            size = "500M";
+            type = "EF00";
             content = {
               type = "filesystem";
               format = "vfat";
               mountpoint = "/boot";
             };
-          }
-          {
-            name = "luks";
-            start = "512MiB";
-            end = "100%";
+          };
+
+          luks = {
+            size = "100%";
             content = {
               type = "luks";
               name = "crypted";
@@ -38,10 +34,11 @@
                 vg = "pool";
               };
             };
-          }
-        ];
+          };
+        };
       };
     };
+
     lvm_vg.pool = {
       type = "lvm_vg";
       lvs = {